relocate toggle action, fix toggle verbage, more generic section args
